# apps/web/src/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalysisResult": {
        "properties": {
          "analysis_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Id"
          },
          "case_quality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseQualitySummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "cross_source_links": {
            "items": {
              "$ref": "#/components/schemas/CrossSourceLink"
            },
            "title": "Cross Source Links",
            "type": "array"
          },
          "entities": {
            "items": {
              "$ref": "#/components/schemas/Entity"
            },
            "title": "Entities",
            "type": "array"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/ParsedEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLine"
            },
            "title": "Evidence",
            "type": "array"
          },
          "evidence_integrity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvidenceIntegritySummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "finding_count": {
            "title": "Finding Count",
            "type": "integer"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "title": "Findings",
            "type": "array"
          },
          "incident_count": {
            "title": "Incident Count",
            "type": "integer"
          },
          "incidents": {
            "items": {
              "$ref": "#/components/schemas/Incident"
            },
            "title": "Incidents",
            "type": "array"
          },
          "live_retention": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LiveRetentionSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "live_snapshot_attestation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Live Snapshot Attestation"
          },
          "parsed_event_count": {
            "title": "Parsed Event Count",
            "type": "integer"
          },
          "parser": {
            "title": "Parser",
            "type": "string"
          },
          "raw_line_count": {
            "title": "Raw Line Count",
            "type": "integer"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceSummary"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "source_id",
          "parser",
          "raw_line_count",
          "parsed_event_count",
          "finding_count",
          "incident_count"
        ],
        "title": "AnalysisResult",
        "type": "object"
      },
...
      "Body_analyze_case_bundle_upload_api_analyze_case_bundle_post": {
        "properties": {
          "files": {
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          }
        },
        "required": [
          "files"
        ],
        "title": "Body_analyze_case_bundle_upload_api_analyze_case_bundle_post",
        "type": "object"
      },
      "Body_analyze_upload_api_analyze_upload_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_analyze_upload_api_analyze_upload_post",
        "type": "object"
      },
      "CaseQualitySummary": {
        "properties": {
          "cross_source_corroborated_incident_count": {
            "default": 0,
            "title": "Cross Source Corroborated Incident Count",
            "type": "integer"
          },
          "sequence_backed_incident_count": {
            "default": 0,
            "title": "Sequence Backed Incident Count",
            "type": "integer"
          },
          "strongest_incident_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strongest Incident Id"
          },
          "strongest_incident_score": {
            "default": 0,
            "title": "Strongest Incident Score",
            "type": "integer"
          },
          "strongest_incident_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strongest Incident Title"
          },
          "top_scoring_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Scoring Reason"
          },
          "total_cross_source_links": {
            "default": 0,
            "title": "Total Cross Source Links",
            "type": "integer"
          }
        },
        "title": "CaseQualitySummary",
        "type": "object"
      },
...
      "CrossSourceLink": {
        "properties": {
          "confidence": {
            "default": "medium",
            "title": "Confidence",
            "type": "string"
          },
          "destination_ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Ip"
          },
          "destination_port": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Port"
          },
          "event_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Time"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "link_type": {
            "title": "Link Type",
            "type": "string"
          },
          "match_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Value"
          },
          "source_event_id": {
            "title": "Source Event Id",
            "type": "string"
          },
          "source_event_type": {
            "title": "Source Event Type",
            "type": "string"
          },
          "source_ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ip"
          },
          "source_label": {
            "title": "Source Label",
            "type": "string"
          },
          "source_raw_line_id": {
            "title": "Source Raw Line Id",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "target_event_id": {
            "title": "Target Event Id",
            "type": "string"
          },
          "target_event_type": {
            "title": "Target Event Type",
            "type": "string"
          },
          "target_label": {
            "title": "Target Label",
            "type": "string"
          },
          "target_raw_line_id": {
            "title": "Target Raw Line Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "link_type",
          "source_event_id",
          "target_event_id",
          "source_raw_line_id",
          "target_raw_line_id",
          "source_label",
          "target_label",
          "source_event_type",
          "target_event_type",
          "summary"
        ],
        "title": "CrossSourceLink",
        "type": "object"
      },
      "Entity": {
        "properties": {
          "analysis_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Id"
          },
          "entity_type": {
            "enum": [
              "ip",
              "user",
              "host",
              "service",
              "path",
              "domain",
              "url",
              "container"
            ],
            "title": "Entity Type",
            "type": "string"
          },
          "event_count": {
            "default": 0,
            "title": "Event Count",
            "type": "integer"
          },
          "finding_ids": {
            "items": {
              "type": "string"
            },
            "title": "Finding Ids",
            "type": "array"
          },
          "first_seen": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Seen"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "incident_ids": {
            "items": {
              "type": "string"
            },
            "title": "Incident Ids",
            "type": "array"
          },
          "last_seen": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          },
          "risk_score": {
            "default": 0,
            "title": "Risk Score",
            "type": "integer"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "title": "Source Ids",
            "type": "array"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "id",
          "entity_type",
          "value"
        ],
        "title": "Entity",
        "type": "object"
      },
      "EvidenceIntegritySummary": {
        "properties": {
          "algorithm": {
            "const": "sha256",
            "default": "sha256",
            "title": "Algorithm",
            "type": "string"
          },
          "attested_live_snapshot": {
            "default": false,
            "title": "Attested Live Snapshot",
            "type": "boolean"
          },
          "live_retention": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LiveRetentionSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "origin": {
            "enum": [
              "legacy",
              "upload",
              "case_bundle",
              "live_snapshot",
              "file_watch",
              "folder_watch",
              "docker_logs",
              "interface_packets",
              "syslog"
            ],
            "title": "Origin",
            "type": "string"
          },
          "status": {
            "default": "verified",
            "enum": [
              "verified",
              "legacy_unverified",
              "raw_purged"
            ],
            "title": "Status",
            "type": "string"
          },
          "verified_line_count": {
            "minimum": 0.0,
            "title": "Verified Line Count",
            "type": "integer"
          }
        },
        "required": [
          "origin",
          "verified_line_count"
        ],
        "title": "EvidenceIntegritySummary",
        "type": "object"
      },
      "EvidenceLine": {
        "properties": {
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "line_number": {
            "title": "Line Number",
            "type": "integer"
          },
          "raw_text": {
            "title": "Raw Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "line_number",
          "raw_text",
          "content_hash"
        ],
        "title": "EvidenceLine",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Incident": {
        "properties": {
          "entities": {
            "items": {
              "type": "string"
            },
            "title": "Entities",
            "type": "array"
          },
          "finding_ids": {
            "items": {
              "type": "string"
            },
            "title": "Finding Ids",
            "type": "array"
          },
          "first_seen": {
            "format": "date-time",
            "title": "First Seen",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_seen": {
            "format": "date-time",
            "title": "Last Seen",
            "type": "string"
          },
          "mitre_techniques": {
            "items": {
              "type": "string"
            },
            "title": "Mitre Techniques",
            "type": "array"
          },
          "score": {
            "title": "Score",
            "type": "integer"
          },
          "score_breakdown": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Score Breakdown",
            "type": "object"
          },
          "score_rationale": {
            "items": {
              "type": "string"
            },
            "title": "Score Rationale",
            "type": "array"
          },
          "severity": {
            "enum": [
              "info",
              "low",
              "medium",
              "high",
              "critical"
            ],
            "title": "Severity",
            "type": "string"
          },
          "status": {
            "default": "active",
            "enum": [
              "active",
              "investigating",
              "closed",
              "false_positive"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "timeline": {
            "items": {
              "type": "string"
            },
            "title": "Timeline",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "severity",
          "summary",
          "first_seen",
          "last_seen",
          "score"
        ],
        "title": "Incident",
        "type": "object"
      },
...
      "ParsedEvent": {
        "properties": {
          "event_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Time"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "normalized_fields": {
            "additionalProperties": true,
            "title": "Normalized Fields",
            "type": "object"
          },
          "raw_line_id": {
            "title": "Raw Line Id",
            "type": "string"
          },
          "service": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ip"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "required": [
          "id",
          "source_id",
          "raw_line_id",
          "event_type",
          "message"
        ],
        "title": "ParsedEvent",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local-only live SOC assistant API.",
    "title": "TraceHawk API",
    "version": "0.9.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/analyze/case-bundle": {
      "post": {
        "operationId": "analyze_case_bundle_upload_api_analyze_case_bundle_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_case_bundle_upload_api_analyze_case_bundle_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze Case Bundle Upload",
        "tags": [
          "analysis"
        ]
      }
    },
...
    "/api/analyze/incidents": {
      "get": {
        "operationId": "recent_incidents_api_analyze_incidents_get",
        "parameters": [
          {
            "in": "query",
            "name": "analysis_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Incident"
                  },
                  "title": "Response Recent Incidents Api Analyze Incidents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recent Incidents",
        "tags": [
          "analysis"
        ]
      }
    },
...
    "/api/analyze/runs/{analysis_id}": {
      "get": {
        "operationId": "analysis_run_detail_api_analyze_runs__analysis_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analysis Run Detail",
        "tags": [
          "analysis"
        ]
      }
    },
    "/api/analyze/sample/{sample_id}": {
      "get": {
        "operationId": "analyze_sample_api_analyze_sample__sample_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "sample_id",
            "required": true,
            "schema": {
              "title": "Sample Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze Sample",
        "tags": [
          "analysis"
        ]
      }
    },
    "/api/analyze/upload": {
      "post": {
        "operationId": "analyze_upload_api_analyze_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_upload_api_analyze_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze Upload",
        "tags": [
          "analysis"
        ]
      }
    },
...
  }
}